refactored code with controller and async middleware
