+++ middleware/async.js
@@ -1,0 +1,5 @@
+const asyncHandler  = (fn) => (req,res,next)=>
+Promise.resolve(fn(req,res,next)).catch(next)
+
+
+module.exports = asyncHandler
